Update with new logic to rewrite this project

+++ src/main/java/com/sushishop/repository/SushiRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.sushishop.entity.Sushi;;
 
 @Repository
-public interface SushiRepository extends JpaRepository<Sushi, Long>{
+public interface SushiRepository extends JpaRepository<Sushi, Integer>{
 
     /**
      * Find sushi by name
